proto2 required fields guard
